+++ packages/react-docs/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.5.0-prerelease.86](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.5.0-prerelease.85...@patternfly/react-docs@7.5.0-prerelease.86) (2026-05-08)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.5.0-prerelease.85](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.5.0-prerelease.84...@patternfly/react-docs@7.5.0-prerelease.85) (2026-05-08)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.5.0-prerelease.84](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.5.0-prerelease.83...@patternfly/react-docs@7.5.0-prerelease.84) (2026-05-08)
+
+### Features
+
+- **Hero:** glass styling update ([#12412](https://github.com/patternfly/patternfly-react/issues/12412)) ([bea138e](https://github.com/patternfly/patternfly-react/commit/bea138ecd4d33a2f591dac628b875b295ec71c31))
+
+# [7.5.0-prerelease.83](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.5.0-prerelease.82...@patternfly/react-docs@7.5.0-prerelease.83) (2026-05-08)
+
+### Features
+
+- **Page:** add dynamic sticky section support ([#12409](https://github.com/patternfly/patternfly-react/issues/12409)) ([97b29d3](https://github.com/patternfly/patternfly-react/commit/97b29d3fafc14ee01f0bbf0d2bdbc527d62f2a16))
+
+# [7.5.0-prerelease.82](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.5.0-prerelease.81...@patternfly/react-docs@7.5.0-prerelease.82) (2026-05-07)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.5.0-prerelease.81](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.5.0-prerelease.80...@patternfly/react-docs@7.5.0-prerelease.81) (2026-05-07)
+
+### Features
+
+- **Pagination:** added dynamic sticky and plain ([#12388](https://github.com/patternfly/patternfly-react/issues/12388)) ([7c4f9df](https://github.com/patternfly/patternfly-react/commit/7c4f9dfac4c4a63deecd9fb4d517c5713a617f9e))
+
+# [7.5.0-prerelease.80](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.5.0-prerelease.79...@patternfly/react-docs@7.5.0-prerelease.80) (2026-05-07)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
 # [7.5.0-prerelease.79](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.5.0-prerelease.78...@patternfly/react-docs@7.5.0-prerelease.79) (2026-05-05)
 
 **Note:** Version bump only for package @patternfly/react-docs